requirements.txt
pillow
numpy
tensorflow
scikit-learn
opencv-python